refactor Entity class to remove interfaces
